newspaper-multicrop2: altered allot, READ DESC
Changed deskew to add extra transparent pixels instead of extra white pixels.  With out this there is a bug where the entire background of the image is changed to white which causes the content of the news clippings to be consumed with flood fill. Read comments in the file. I did not update newspaper-multicrop yet, I only updated newspaper-multicrop2.